refactor: rename `Models` dir to `Clauses`
