+++ src/Framework/QueryBuilder/Clauses/RawSQL.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Give\Framework\QueryBuilder\Models;
+namespace Give\Framework\QueryBuilder\Clauses;
 
 use Give\Framework\Database\DB;
