+++ src/Framework/QueryBuilder/Clauses/Select.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Give\Framework\QueryBuilder\Models;
+namespace Give\Framework\QueryBuilder\Clauses;
 
 /**
  * @unreleased
